requirements.txt
tensorflow-gpu==2.0.0
flake8
pytest
h5py
PyYAML
ipython
